pyproject.toml
[project]
name = "multi-agent-research-orchestration-framework"
version = "0.2.0"
description = "Automated agentic research system with arXiv graph intelligence and Discord reporting"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.40",
    "ddgs>=9.0",
    "tavily-python>=0.5",
    "httpx>=0.27",
    "click>=8.1",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "rich>=13.0",
    # arXiv research graph pipeline
    "arxiv>=2.1",
    "pymupdf>=1.24",
    "numpy>=1.26",
    # PDF generation & S3 upload
    "boto3>=1.34",
    "weasyprint>=62.0",
]

[project.scripts]
research = "daily_research.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/daily_research"]
